GenAI_AI_Resume_Screening_System_with_Tracing_Task_3/requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
langsmith>=0.1.130
python-dotenv>=1.0.1
